move shared types and fw rev parsing into a mwocp6x module
