+++ drv/i2c-devices/src/lib.rs
@@ -288,8 +288,7 @@ pub mod max31790;
 pub mod max5970;
 pub mod max6634;
 pub mod mcp9808;
-pub mod mwocp67;
-pub mod mwocp68;
+pub mod mwocp6x;
 pub mod nvme_bmc;
 pub mod pca9538;
 pub mod pca9956b;
