+++ task/power/src/main.rs
@@ -17,8 +17,7 @@ use drv_i2c_devices::lm5066::*;
 use drv_i2c_devices::lm5066i::*;
 use drv_i2c_devices::ltc4282::*;
 use drv_i2c_devices::max5970::*;
-use drv_i2c_devices::mwocp67::*;
-use drv_i2c_devices::mwocp68::*;
+use drv_i2c_devices::mwocp6x::*;
 use drv_i2c_devices::raa229618::*;
 use drv_i2c_devices::raa229620a::*;
 use drv_i2c_devices::tps546b24a::*;
